backend/requirements.txt
fastapi==0.111.0
pydantic==2.7.4
uvicorn==0.30.1
PyJWT==2.8.0
cryptography==42.0.8
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
requests==2.32.3
beautifulsoup4==4.12.3
APScheduler==3.10.4
